validate

+++ src/scripts/updateEvent.js
@@ -10,15 +10,18 @@ const updateEvent = (event) => {
     const endTime = document.querySelector('.end-time')
     let inputDescription = document.querySelector('.description-input');
 
+    let strat = new Date(startDate.value + 'T' + startTime.value);
+    let end = new Date(endDate.value + 'T' + endTime.value)
+    if (!duration(strat, end)) return;
+    if (!errorDate(strat.getTime(), end.getTime())) return;
+
+
     deleteAll()
 
     selectedElem.name = inputName.value
     selectedElem.startDateEvent = startDate.value + 'T' + startTime.value;
     selectedElem.endDateEvent = endDate.value + 'T' + endTime.value;
     selectedElem.description = inputDescription.value;
-
-
-
 
     renderEvents()
 }
